requirements.txt
setuptools
wheel
mlflow==1.9.1
scikit-learn==0.23.2
pytest==5.3.1
python-dotenv==0.10.3
databricks-cli==0.10.0
pyyaml>=5.1
click==7.1.2
cookiecutter==1.7.2